requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.11.0
FlagEmbedding==1.2.11
openai==1.37.0
langchain-text-splitters==0.2.2
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0

# Phase 3: User System
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.9.0
